% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{bdiagMatrix}
\alias{bdiagMatrix}
\title{Create a block diagonal matrix}
\usage{
bdiagMatrix(mlist, fill = 0)
}
\arguments{
\item{mlist}{a list of matrices}

\item{fill}{value to fill the non-diagnoal elements}
}
\value{
a block diagonal matrix
}
\description{
Create a block diagonal matrix
}
\examples{

m1 <- bdiagMatrix(list(matrix(1:4, 2), diag(3)))
m2 <- bdiagMatrix(list(matrix(1:4, 2), diag(3)),fill = NA)

}
\seealso{
\code{bdiag} in \pkg{Matrix}
}
