% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedUtils.R
\name{cjoint}
\alias{cjoint}
\title{Combine R Objects by Columns}
\usage{
cjoint(..., fill = NA)
}
\arguments{
\item{...}{vectors or matrices}

\item{fill}{a scalar used when these objects have different number of rows.}
}
\value{
a data frame
}
\description{
Combine a sequence of vector, matrix or data-frame arguments by columns. Vector is treated as a column matrix.
}
\examples{
cjoint(2,c(1,2,3,4),matrix(1:6,2,3))
cjoint(v1 = 2, v2 = c(3,2), v3 = matrix(1:6,3,2),
       v4 = data.frame(c(3,4,5,6,7),rep("x",5)),fill = 99)

}
\seealso{
\code{\link[base]{cbind}}
}

