% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM.R
\name{DTM}
\alias{DTM}
\title{Calibrate diagnostic tree model (experimental)}
\usage{
DTM(dat, Qc, delta = NULL, Tmatrix = NULL, conv.crit = 0.001,
  conv.type = "pr", maxitr = 1000)
}
\arguments{
\item{dat}{A required \eqn{N \times J} data matrix of N examinees to J items. Missing
values are currently not allowed.}

\item{Qc}{A required \eqn{J \times K+2} category and attribute association matrix,
where J represents the number of items or nonzero categories and K represents the
number of attributes. Entry 1 indicates that the attribute is
measured by the item, and 0 otherwise. The first column gives the item number, which must
be numeric and match the number of column in the data. The second column indicates the category number.}

\item{delta}{initial item parameters}

\item{Tmatrix}{The mapping matrix showing the relation between the OBSERVED responses (rows) and the PSEDUO items (columns);
The first column gives the observed responses.}

\item{conv.crit}{The convergence criterion for max absolute change in item parameters.}

\item{conv.type}{convergence criteria; Can be \code{pr},\code{LL} and \code{delta},
indicating category response function, log-likelihood and delta parameters,respectively.}

\item{maxitr}{The maximum iterations allowed.}
}
\description{
This function estimates the diagnostic tree model (Ma, 2018). It is an experimental function, and it is expected to be further extended.
}
\references{
Ma, W. (2018). A Diagnostic Tree Model for Polytomous Responses with Multiple Strategies. \emph{British Journal of Mathematical and Statistical Psychology.}
}
