% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.itemfit.R
\name{heatplot}
\alias{heatplot}
\title{Item fit plots}
\usage{
heatplot(object, ...)
}
\arguments{
\item{object}{model object of class \code{itemfit}}

\item{...}{additional arguments}
}
\description{
Create plots of bivariate heatmap for item fit
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ

fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
ift <- itemfit(fit)
heatplot(ift)
}
}
\seealso{
\code{\link{GDINA}}, \code{\link{itemfit}}
}
