% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoparm.R
\name{hoparm}
\alias{hoparm}
\title{extract higher-order parameters}
\usage{
hoparm(object, withSE = FALSE, digits = 4, ...)
}
\arguments{
\item{object}{estimated GDINA object returned from \code{\link{GDINA}}}

\item{withSE}{estimate standard errors for lambda parameters or not?}

\item{digits}{how many decimal places for the ouput?}

\item{...}{additional arguments}
}
\value{
a list with element \code{theta} for higher-order incidental (ability) parameters
and \code{lambda} for higher-order structural parameters.
}
\description{
Function to extract higher-order parameters when a higher-order model is fitted.
}
\author{
{Wenchao Ma, Rutgers University, \email{wenchao.ma@rutgers.edu} \cr Jimmy de la Torre, The University of Hong Kong}
}

