% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILCA.R
\name{ILCA}
\alias{ILCA}
\title{Iterative latent-class analysis}
\usage{
ILCA(dat, Q, seed.num = 5)
}
\arguments{
\item{dat}{A required binary item response matrix.}

\item{Q}{A required binary item and attribute association matrix.}

\item{seed.num}{seed number; Default = 5.}
}
\value{
Estimated attribute profiles.
}
\description{
This function implements an iterative latent class analysis (ILCA; Jiang, 2019) approach to estimating attributes for cognitive diagnosis.
}
\examples{
\dontrun{
ILCA(sim10GDINA$simdat, sim10GDINA$simQ)
}
}
\references{
Jiang, Z. (2019). Using the iterative latent-class analysis approach to improve attribute accuracy in diagnostic classification models. \emph{Behavior research methods}, 1-10.
}
\author{
{Zhehan Jiang, The University of Alabama}
}
