% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3GDINA.R
\name{indlogLik}
\alias{indlogLik}
\title{Extract log-likelihood for each individual}
\usage{
indlogLik(object, ...)
}
\arguments{
\item{object}{GDINA object}

\item{...}{additional arguments}
}
\description{
Extract individual log-likelihood.
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ

fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
iL <- indlogLik(fit)
iL[1:6,]
}
}
