% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3GDINA.R
\name{indlogPost}
\alias{indlogPost}
\title{Extract log posterior for each individual}
\usage{
indlogPost(object, ...)
}
\arguments{
\item{object}{GDINA object}

\item{...}{additional arguments}
}
\description{
Extract individual log posterior.
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
iP <- indlogPost(fit)
iP[1:6,]
}
}
