% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{init_design_matrices}
\alias{init_design_matrices}
\title{Initialize design matrices}
\usage{
init_design_matrices(
  ncat,
  model,
  rule,
  Kj,
  reduced.LG,
  Q,
  originalQ,
  no.bugs = NULL
)
}
\arguments{
\item{ncat}{Number of categories}

\item{model}{Vector of model codes}

\item{rule}{Vector of condensation rules}

\item{Kj}{Vector of attribute counts per item}

\item{reduced.LG}{List of reduced latent group matrices}

\item{Q}{Q-matrix (attributes only, no item/category columns)}

\item{originalQ}{Original Q-matrix with item/category columns for MS-DINA}

\item{no.bugs}{Number of bugs for BUGDINO model}
}
\value{
List of design matrices
}
\description{
Creates design matrices for all items based on model type and Q-matrix.
}
\keyword{internal}
