% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{init_solver}
\alias{init_solver}
\title{Initialize solver vector}
\usage{
init_solver(solver, ncat)
}
\arguments{
\item{solver}{Character string or vector specifying solver}

\item{ncat}{Number of categories}
}
\value{
Character vector of length ncat
}
\description{
Expands solver specification to per-item vector.
}
\keyword{internal}
