% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{init_solver_args}
\alias{init_solver_args}
\title{Initialize solver arguments}
\usage{
init_solver_args(auglag_args, solnp_args, nloptr_args)
}
\arguments{
\item{auglag_args}{User-provided auglag arguments}

\item{solnp_args}{User-provided solnp arguments}

\item{nloptr_args}{User-provided nloptr arguments}
}
\value{
List containing initialized solver arguments
}
\description{
Sets up default solver parameters for nloptr-based optimization.
}
\keyword{internal}
