% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{internal_GDINA}
\alias{internal_GDINA}
\alias{internal_Lik}
\alias{internal_aggregateCol}
\alias{internal_eta}
\alias{internal_matchMatrix}
\alias{internal_RowNormalize}
\alias{internal_ColNormalize}
\alias{internal_RowProd}
\alias{internal_l2m}
\alias{internal_l2m}
\alias{internal_m2l}
\alias{internal_uP}
\title{Functions for internal use}
\usage{
internal_Lik(...)

internal_aggregateCol(...)

internal_eta(...)

internal_matchMatrix(...)

internal_RowNormalize(...)

internal_ColNormalize(...)

internal_RowProd(...)

internal_l2m(...)

internal_m2l(...)

internal_uP(...)
}
\arguments{
\item{...}{arguments passed to internal functions}
}
\description{
Functions for internal use
}
