% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemparm.GDINA.R
\name{itemparm.GDINA}
\alias{itemparm.GDINA}
\title{extract lower-order structural (item) parameters}
\usage{
\method{itemparm}{GDINA}(object, what = c("catprob", "itemprob", "LCprob",
  "gs", "delta", "rrum"), withSE = FALSE, SE.type = 2, digits = 4, ...)
}
\arguments{
\item{object}{estimated GDINA object for various S3 methods}

\item{what}{argument for various S3 methods}

\item{withSE}{argument for method \code{\link{itemparm}}; show standard errors or not?}

\item{SE.type}{type of standard errors.}

\item{digits}{How many decimal places in each number? The default is 4.}

\item{...}{additional arguments}
}
\description{
To extract lower-order structural (item) parameters, use method \code{\link{itemparm}}.
}
