% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3GDINA.R
\name{npar}
\alias{npar}
\title{Calculate the number of parameters}
\usage{
npar(object, ...)
}
\arguments{
\item{object}{GDINA object}

\item{...}{additional arguments}
}
\description{
Calculate the number of parameters for GDINA estimates.
Returned the total number of parameters, the number of item parameters and
the number parameters of joint attribute distribution.
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ

fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
npar(fit)
}
}
