% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRF.GDINA.R
\name{plot.itemfit}
\alias{plot.itemfit}
\title{Item fit plots}
\usage{
\method{plot}{itemfit}(x, ...)
}
\arguments{
\item{x}{model object of class \code{itemfit}}

\item{...}{additional arguments}
}
\description{
Create plots of bivariate heatmap for item fit
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ

fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
ift <- itemfit(fit)
plot(ift)
}
}
\seealso{
\code{\link{GDINA}}, \code{\link{itemfit}}
}
