% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRF.GDINA.R
\name{plot.pairwiseDIF}
\alias{plot.pairwiseDIF}
\title{Grouped bar plots for pairwise DIF posthoc analysis}
\usage{
\method{plot}{pairwiseDIF}(x, item = "all", withSE = FALSE, SE.type = 2, ...)
}
\arguments{
\item{x}{model object of class \code{\link{pairwiseDIF}}}

\item{item}{A scalar or vector specifying the DIF item(s) to plot.}

\item{withSE}{logical; Add error bar (estimate - SE, estimate + SE) to the plots?}

\item{SE.type}{How is SE estimated. By default, it's based on OPG using incomplete information.}

\item{...}{additional arguments}
}
\description{
Create grouped bar charts of group-specific item/category response probabilities
for DIF items identified by \code{pairwiseDIF()}.
}
\seealso{
\code{\link{pairwiseDIF}}, \code{\link{dif}}
}
