% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_utils.R
\name{print_em_progress}
\alias{print_em_progress}
\title{Print EM iteration progress}
\usage{
print_em_progress(itr, maxchg, neg2LL, verbose)
}
\arguments{
\item{itr}{Iteration number}

\item{maxchg}{Maximum change}

\item{neg2LL}{Current deviance}

\item{verbose}{Verbosity level (1 = inline, 2 = newline)}
}
\description{
Prints iteration information with appropriate formatting.
}
\keyword{internal}
