% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCount.R
\name{rowCount}
\alias{rowCount}
\title{Count and label unique rows in data frame}
\usage{
rowCount(df)
}
\arguments{
\item{df}{a data frame or matrix}
}
\value{
freq the number of rows

group the data frame with a column named row.no giving unique labels for all unique rows
}
\description{
Count and label unique rows in data frame
}
\examples{

df <- data.frame(V1=c(1L,2L),V2=LETTERS[1:3],V3=rep(1,12))
rowCount(df)



}

