% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{rowMatch}
\alias{rowMatch}
\title{Count the frequency of a row vector in a data frame}
\usage{
rowMatch(df, vec = NULL)
}
\arguments{
\item{df}{a data frame or matrix}

\item{vec}{the vector for matching}
}
\value{
count the number of vector vec in the data frame

row.no row numbers of the vector vec in the data frame
}
\description{
Count the frequency of a row vector in a data frame
}
\examples{

df <- data.frame(V1=c(1L,2L),V2=LETTERS[1:3],V3=rep(1,12))
rowMatch(df,c(2,"B",1))



}
