% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{score}
\alias{score}
\title{Score function}
\usage{
score(object, parm = "delta")
}
\arguments{
\item{object}{an object of class GDINA}

\item{parm}{Either \code{delta} or \code{prob} indicating score function for delta parameters and
success probabily parameters}
}
\value{
a list where elements give the score functions for each item or category
}
\description{
Calculate score function for each dichotomous item or each nonzero category for polytomous items
Only applicable to saturated model ofr joint attribute distribution
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
score(fit)
}


}
