% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim10GDINA.R
\docType{data}
\name{sim10GDINA}
\alias{sim10GDINA}
\title{Simulated data, Q-matrix and item parameters (10 items, G-DINA model)}
\format{A list of Q-matrix and item parameters for a 10-item test with components:
\describe{
\item{\code{simdat}}{simulated responses of 1000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
\item{\code{simItempar}}{artificial item parameters (probability of success for each latent group)}
}}
\usage{
sim10GDINA
}
\description{
Artificial Q-matrix and item parameters for a 10-item test measuring 3 attributes
are given. Based on the Q-matrix and item parameters, responses of 1000 examinees
were simulated using the random seed of 12345.
}
\keyword{datasets}
