% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim10MCDINA2.R
\docType{data}
\name{sim10MCDINA2}
\alias{sim10MCDINA2}
\title{Simulated data (10 items, MC-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 3000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
}
}
\usage{
sim10MCDINA2
}
\description{
Simulated data, Q-matrix and item parameters for a 10-item test measuring 5 attributes.
}
\references{
Ma, W., & de la Torre, J. (2020). GDINA: An R Package for Cognitive Diagnosis Modeling. \emph{Journal of Statistical Software, 93(14)}, 1-26.
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
\keyword{datasets}
