% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim20seqGDINA.R
\docType{data}
\name{sim20seqGDINA}
\alias{sim20seqGDINA}
\title{Simulated data (20 items, sequential G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated polytomous responses of 2000 examinees}
\item{\code{simQ}}{artificial Qc-matrix}
\item{\code{simItempar}}{artificial item parameters (category level probability of success for each latent group)}
}
}
\usage{
sim20seqGDINA
}
\description{
Simulated data, Qc-matrix and item parameters for a 20-item test measuring 5 attributes.
}
\references{
Ma, W., & de la Torre, J. (2020). GDINA: An R Package for Cognitive Diagnosis Modeling. \emph{Journal of Statistical Software, 93(14)}, 1-26.
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
\keyword{datasets}
