% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim30GDINA.R
\docType{data}
\name{sim30GDINA}
\alias{sim30GDINA}
\title{Simulated data (30 items, G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 1000 examinees}
\item{\code{simQ}}{artificial \eqn{30 \times 5} Q-matrix}
\item{\code{simItempar}}{artificial item parameters(probability of success for each latent group)}
}
}
\usage{
sim30GDINA
}
\description{
Simulated data, Q-matrix and item parameters for a 30-item test measuring 5 attributes.
}
\references{
Ma, W., & de la Torre, J. (2020). GDINA: An R Package for Cognitive Diagnosis Modeling. \emph{Journal of Statistical Software, 93(14)}, 1-26.
}
\author{
{Wenchao Ma, The University of Alabama, \email{wenchao.ma@ua.edu}}
}
\keyword{datasets}
