% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim30pGDINA.R
\docType{data}
\name{sim30pGDINA}
\alias{sim30pGDINA}
\title{Simulated data (30 items, polytomous G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 3000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
\item{\code{simItempar}}{artificial item parameters(probability of success for each latent group)}
}
}
\usage{
sim30pGDINA
}
\description{
Simulated data, Q-matrix and item parameters for a 30-item test measuring 5 attributes.
}
\keyword{datasets}
