% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGDINA.R
\name{startGDINA}
\alias{startGDINA}
\title{Graphical user interface of the GDINA function}
\usage{
startGDINA()
}
\description{
An experimental interactive Shiny application for running GDINA function
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
startGDINA()
}

}

