% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{unique_only}
\alias{unique_only}
\title{Unique values in a vector}
\usage{
unique_only(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
sorted unique values
}
\description{
Unique values in a vector
}
\examples{

vec <- c(4,2,3,5,4,4,4)
unique_only(vec)
# see the difference from unique
unique(vec)

vec <- letters[1:5]
unique_only(vec)



}
\seealso{
\link[base]{unique}
}
