% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportedFuncs.R
\name{unrestrQ}
\alias{unrestrQ}
\title{Generate unrestricted Qc matrix from an restricted Qc matrix}
\usage{
unrestrQ(Qc)
}
\arguments{
\item{Qc}{an restricted Qc matrix}
}
\value{
an unrestricted Qc matrix
}
\description{
Generate unrestricted Qc matrix from an restricted Qc matrix
}
\examples{

Qc <- sim21seqDINA$simQc
Qc
unrestrQ(Qc)


}
