% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDINA-package.R
\docType{package}
\name{GDINA-package}
\alias{GDINA-package}
\title{The Generalized DINA Model Framework}
\description{
For conducting CDM analysis within the G-DINA model framework
}
\details{
This package provides a framework for a series of cognitively diagnostic analyses
for dichotomous and polytomous responses.

Various cognitive
diagnosis models (CDMs) can be calibrated using the \code{\link{GDINA}}
function, including the G-DINA model (de la Torre, 2011), the deterministic inputs,
noisy and gate (DINA; de la Torre, 2009; Junker & Sijtsma, 2001) model,
the deterministic inputs, noisy or gate (DINO; Templin & Henson, 2006)
model, the reduced reparametrized unified model (R-RUM; Hartz, 2002),
the additive CDM (A-CDM; de la Torre, 2011), and the linear logistic
model (LLM; Maris, 1999), the multiple-strategy DINA model (de la Torre, & Douglas, 2008) and models defined
by users under the G-DINA framework using different link functions and design
matrices (de la Torre, 2011). Note that the LLM is also called
compensatory RUM and the RRUM is equivalent to the generalized NIDA model.

For ordinal and nominal responses,
the sequential G-DINA model (Ma, & de la Torre, 2016) can be fitted and most of the
aforementioned CDMs can be used as the processing functions (Ma, & de la Torre, 2016) at the category level.
Different CDMs can be assigned to different items within a single assessment.
Item parameters are estimated using the MMLE/EM algorithm. Details about the estimation algorithm
can be found in de la Torre (2009), de la Torre (2011),
Ma, Iaconangelo, & de la Torre (2016) and Ma, & de la Torre (2016).
The joint attribute distribution can be modelled using an independent model,
a higher-order IRT model (de la Torre, & Douglas, 2004), a loglinear model (Xu & von Davier, 2008),
a saturated model or a hierarchical structures (e.g., linear, divergent). Monotonicity constraints for item/category success
probabilities can also be specified.

In addition, to handle multiple strategies in ordinal response data, diagnostic treee model (Ma, 2018) can also
be estimated using \code{\link{DTM}} function. Note that this function is experimental, and is expected to be further extended
in the future.

Q-matrix validation (de la Torre, & Chiu, 2016; see \code{\link{Qval}}),
imodel fit statistics (see \code{\link{modelfit}} and \code{\link{itemfit}}),
model comparison at test and item level (de la Torre, & Lee, 2013;
Ma, Iaconangelo, & de la Torre, 2016; see \code{\link{modelcomp}}),
and differential item functioning (Hou, de la Torre, & Nandakumar, 2014;
see \code{\link{dif}}) can also be conducted.
}
\references{
Chen, J., & de la Torre, J. (2013). A General Cognitive Diagnosis Model for Expert-Defined Polytomous Attributes.
\emph{Applied Psychological Measurement, 37}, 419-437.

Chen, J., de la Torre, J., & Zhang, Z. (2013). Relative and Absolute Fit Evaluation in Cognitive Diagnosis Modeling.
\emph{Journal of Educational Measurement, 50}, 123-140.

de la Torre, J. (2009). DINA Model and Parameter Estimation: A Didactic. \emph{Journal of Educational and Behavioral Statistics, 34}, 115-130.

de la Torre, J. (2011). The generalized DINA model framework. \emph{Psychometrika, 76}, 179-199.

de la Torre, J. & Chiu, C-Y. (2016). A General Method of Empirical Q-matrix Validation. \emph{Psychometrika, 81}, 253-273.

de la Torre, J., & Douglas, J. A. (2004). Higher-order latent trait models for cognitive diagnosis.
\emph{Psychometrika, 69}, 333-353.

de La Torre, J., & Douglas, J. A. (2008). Model evaluation and multiple strategies in cognitive diagnosis: An analysis of fraction subtraction data. \emph{Psychometrika, 73}, 595.

de la Torre, J., & Lee, Y. S. (2013). Evaluating the wald test for item-level comparison of
saturated and reduced models in cognitive diagnosis. \emph{Journal of Educational Measurement, 50}, 355-373.

Haertel, E. H. (1989). Using restricted latent class models to map the skill structure of achievement items.
\emph{Journal of Educational Measurement, 26}, 301-321.

Hartz, S. M. (2002). A bayesian framework for the unified model for assessing cognitive abilities:
Blending theory with practicality (Unpublished doctoral dissertation). University of Illinois at Urbana-Champaign.

Hou, L., de la Torre, J., & Nandakumar, R. (2014). Differential item functioning assessment in cognitive diagnostic modeling: Application of the Wald test to
investigate DIF in the DINA model. \emph{Journal of Educational Measurement, 51}, 98-125.

Junker, B. W., & Sijtsma, K. (2001). Cognitive assessment models with few assumptions, and connections with nonparametric
item response theory. \emph{Applied Psychological Measurement, 25}, 258-272.

Ma, W. (2018). A Diagnostic Tree Model for Polytomous Responses with Multiple Strategies. \emph{British Journal of Mathematical and Statistical Psychology.}

Ma, W., & de la Torre, J. (2016). A sequential cognitive diagnosis model for polytomous responses. \emph{British Journal of Mathematical and Statistical Psychology. 69,} 253-275.

Ma, W., Iaconangelo, C., & de la Torre, J. (2016). Model similarity, model selection and attribute classification.
\emph{Applied Psychological Measurement, 40}, 200-217.

Maris, E. (1999). Estimating multiple classification latent class models. \emph{Psychometrika, 64}, 187-212.

Xu, X., & von Davier, M. (2008). Fitting the structured general diagnostic model to NAEP data. ETS research report, RR-08-27.
}
\seealso{
\pkg{CDM} for estimating G-DINA model and a set of other CDMs;
\pkg{ACTCD} and \pkg{NPCD}
for nonparametric CDMs; \pkg{dina} for DINA model in Bayesian framework
}
\author{
{Wenchao Ma, The University of Alabama, \email{wenchao.ma@ua.edu} \cr Jimmy de la Torre, The University of Hong Kong}
}
\keyword{package}
