% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.itemfit.R
\name{heatplot}
\alias{heatplot}
\title{Item fit plots}
\usage{
heatplot(object, ...)
}
\arguments{
\item{object}{model object of class \code{itemfit}}

\item{...}{additional arguments}
}
\description{
Create plots of bivariate heatmap for item fit
}
\seealso{
\code{\link{GDINA}}, \code{\link{itemfit}}
}
