% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemparm.GDINA.R
\name{itemparm}
\alias{itemparm}
\alias{itemparm.GDINA}
\title{extract item parameters (deprecated)}
\usage{
itemparm(object, what = c("catprob", "gs", "delta", "rrum", "itemprob",
  "LCprob"), withSE = FALSE, SE.type = 2, digits = 4, ...)

\method{itemparm}{GDINA}(object, what = c("catprob", "gs", "delta", "rrum",
  "itemprob", "LCprob"), withSE = FALSE, SE.type = 2, digits = 4, ...)
}
\arguments{
\item{object}{estimated GDINA object returned from \code{\link{GDINA}}}

\item{what}{what to show.}

\item{withSE}{show standard errors or not?}

\item{SE.type}{Type of standard errors.}

\item{digits}{how many decimal places for the ouput?}

\item{...}{additional arguments}
}
\description{
This function has been deprecated; use \code{coef} instead.
}
\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
# deprecated
itemparm(fit)
coef(fit)
}

}
\references{
Philipp, M., Strobl, C., de la Torre, J., & Zeileis, A.(2017). On the estimation of standard errors in cognitive diagnosis models. \emph{Journal of Educational and Behavioral Statistics, 43}, 88-115.
}
