% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcomp.R, R/extract.R, R/summary.GDINA.R
\name{modelcomp}
\alias{modelcomp}
\alias{extract.modelcomp}
\alias{summary.modelcomp}
\title{Item-level model comparison using Wald, LR or LM tests}
\usage{
modelcomp(GDINA.obj = NULL, method = "Wald", items = "all",
  models = c("DINA", "DINO", "ACDM", "LLM", "RRUM"), DS = FALSE,
  Wald.args = list(SE.type = 2, varcov = NULL), LR.args = list(LR.approx =
  FALSE), LM.args = list(reducedMDINA = NULL, reducedMDINO = NULL,
  reducedMACDM = NULL, SE.type = 2))

\method{extract}{modelcomp}(object, what = c("stats", "pvalues", "df", "DS",
  "models"), digits = 4, ...)

\method{summary}{modelcomp}(object, ...)
}
\arguments{
\item{GDINA.obj}{An estimated model object of class \code{GDINA}}

\item{method}{method for item level model comparison; can be \code{wald}, \code{LR} or \code{LM}.}

\item{items}{a vector of items to specify the items for model comparsion}

\item{models}{a vector specifying which reduced CDMs are possible reduced CDMs for each
item. The default is "DINA","DINO","ACDM","LLM",and "RRUM".}

\item{DS}{whether dissimilarity index should be calculated? \code{FALSE} is the default.}

\item{Wald.args}{a list of options for Wald test including (1) \code{SE.type} giving the type of
covariance matrix for the Wald test; by default, it uses outer product of gradient based on incomplete information matrix;
(2) \code{varcov} for user specified variance-covariance matrix. If supplied, it must
be a list, giving the variance covariance matrix of success probability for each item or category.
The default is \code{NULL}, in which case, the estimated variance-covariance matrix from the GDINA function
is used.}

\item{LR.args}{a list of options for LR test including for now only \code{LR.approx}, which is either \code{TRUE} or \code{FALSE},
indicating whether a two-step LR approximation is implemented or not.}

\item{LM.args}{a list of options for LM test including \code{reducedMDINA}, \code{reducedMDINO}, and \code{reducedMACDM} for
DINA, DINO and ACDM estimates from the \code{GDINA} function; \code{SE.type} specifies the type of covariance matrix.}

\item{object}{object of class \code{modelcomp} for various S3 methods}

\item{what}{argument for S3 method \code{extract} indicating what to extract;
It can be \code{"wald"} for wald statistics, \code{"wald.p"} for associated p-values,
\code{"df"} for degrees of freedom,
and \code{"DS"} for dissimilarity between G-DINA and other CDMs.}

\item{digits}{How many decimal places in each number? The default is 4.}

\item{...}{additional arguments}
}
\value{
an object of class \code{modelcomp}. Elements that can be
extracted using \code{extract} method include
\describe{
\item{stats}{Wald or LR statistics}
\item{pvalues}{p-values associated with the test statistics}
\item{df}{degrees of freedom}
\item{DS}{dissimilarity between G-DINA and other CDMs}
}
}
\description{
This function evaluates whether the saturated G-DINA model can be replaced by reduced
CDMs without significant loss in model data fit for each item using the Wald test, likelihood ratio (LR) test or Lagrange multiplier (LM) test.
For Wald test, see de la Torre and Lee (2013), and Ma, Iaconangelo and de la Torre (2016) for details.
For LR test and a two-step LR approximation procedure, see Sorrel, de la Torre, Abad, and Olea (2017) and Ma (2017).
For LM test, which is only applicable for DINA, DINO and ACDM, see Sorrel, de la Torre, Abad, and Olea (2017) as well. This function also calculates the dissimilarity
between the reduced models and the G-DINA model, which can be viewed as a measure of effect size (Ma, Iaconangelo & de la Torre, 2016).
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: extract various elements from \code{modelcomp} objects

\item \code{summary}: print summary information
}}

\examples{
\dontrun{
dat <- sim10GDINA$simdat
Q <- sim10GDINA$simQ
# --- GDINA model ---#
fit <- GDINA(dat = dat, Q = Q, model = "GDINA")
fit
###################
#
# Wald test
#
###################
w <- modelcomp(fit)
w
# wald statistics
extract(w,"stats")
#p values
extract(w,"pvalues")
##########################
#
# LR and Two-step LR test
#
##########################
lr <- modelcomp(fit,models = c("DINA","DINO"),method = "LR")
lr
TwostepLR <- modelcomp(fit,items =c(6:10),method = "LR",LR.args = list(LR.approx = TRUE))
TwostepLR

##########################
#
# LM test
#
##########################
dina <- GDINA(dat = dat, Q = Q, model = "DINA")
dino <- GDINA(dat = dat, Q = Q, model = "DINO")
acdm <- GDINA(dat = dat, Q = Q, model = "ACDM")
lm <- modelcomp(method = "LM",LM.args=list(reducedMDINA = dina,
reducedMDINO = dino, reducedMACDM  = acdm))
lm


}
}
\references{
de la Torre, J., & Lee, Y. S. (2013). Evaluating the wald test for item-level comparison of
saturated and reduced models in cognitive diagnosis. \emph{Journal of Educational Measurement, 50}, 355-373.

Ma, W., Iaconangelo, C., & de la Torre, J. (2016). Model similarity, model selection and attribute classification.
\emph{Applied Psychological Measurement, 40}, 200-217.

Ma, W. (2017). \emph{A Sequential Cognitive Diagnosis Model for Graded Response: Model Development, Q-Matrix Validation,and Model Comparison. Unpublished doctoral dissertation.} New Brunswick, NJ: Rutgers University.

Sorrel, M. A., de la Torre, J., Abad, F. J., & Olea, J. (2017). Two-Step Likelihood Ratio Test for Item-Level Model Comparison in Cognitive Diagnosis Models. \emph{Methodology, 13}, 39-47.
}
\seealso{
\code{\link{GDINA}}, \code{\link{autoGDINA}}
}
\author{
{Wenchao Ma, The University of Alabama, \email{wenchao.ma@ua.edu} \cr Miguel A. Sorrel, Universidad Autonoma de Madrid \cr Jimmy de la Torre, The University of Hong Kong}
}
