% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim10GDINA.R
\docType{data}
\name{sim10GDINA}
\alias{sim10GDINA}
\title{Simulated data (10 items, G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 1000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
\item{\code{simItempar}}{artificial item parameters (probability of success for each latent group)}
}
}
\usage{
sim10GDINA
}
\description{
Simulated data, Q-matrix and item parameters for a 10-item test with 3 attributes.
}
\references{
Ma, W., & de la Torre, J. (2020). GDINA: An R Package for Cognitive Diagnosis Modeling. \emph{Journal of Statistical Software, 93(14)}, 1-26.
}
\author{
{Wenchao Ma, The University of Alabama, \email{wenchao.ma@ua.edu}}
}
\keyword{datasets}
