% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim20seqGDINA.R
\docType{data}
\name{sim20seqGDINA}
\alias{sim20seqGDINA}
\title{Simulated data, Q-matrix and item parameters (20 items, sequential DINA model)}
\format{A list of Q-matrix and item parameters for a 20-item test with components:
\describe{
\item{\code{simdat}}{simulated polytomous responses of 2000 examinees}
\item{\code{simQ}}{artificial Qc-matrix}
\item{\code{simItempar}}{artificial item parameters(category level probability of success for each latent group)}
}}
\usage{
sim20seqGDINA
}
\description{
Artificial Qc-matrix and item parameters for a 20-item test measuring 5 attributes
are given. Based on the Qc-matrix and item parameters, responses of 2000 examinees
were simulated using the random seed of 12345.
}
\keyword{datasets}

