% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim21seqDINA.R
\docType{data}
\name{sim21seqDINA}
\alias{sim21seqDINA}
\title{Simulated data and Qc-matrix based on the sequential DINA model}
\format{A list of Qc-matrix and responses for a 30-item test with components:
\describe{
\item{\code{simdat}}{simulated responses of 2000 examinees}
\item{\code{simQ}}{artificial Qc-matrix}
}}
\usage{
sim21seqDINA
}
\description{
Artificial Qc-matrix for a 21-item test measuring 5 attributes
are given. Based on the Q-matrix, responses of 2000 examinees
were simulated.
}
\keyword{datasets}

