% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim21seqDINA.R
\docType{data}
\name{sim21seqDINA}
\alias{sim21seqDINA}
\title{Simulated data (21 items, sequential DINA model)}
\format{A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 2000 examinees}
\item{\code{simQ}}{artificial Qc-matrix}
}}
\usage{
sim21seqDINA
}
\description{
Simulated data, and Qc-matrix for a 21-item test measuring 5 attributes.
}
\keyword{datasets}
