% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim30DINA.R
\docType{data}
\name{sim30DINA}
\alias{sim30DINA}
\title{Simulated data (30 items, DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{simulated responses of 1000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
\item{\code{simItempar}}{artificial item parameters (probability of success for each latent group)}
}
}
\usage{
sim30DINA
}
\description{
Simulated data, Q-matrix and item parameters for a 30-item test measuring 5 attributes.
}
\references{
Ma, W., & de la Torre, J. (2020). GDINA: An R Package for Cognitive Diagnosis Modeling. \emph{Journal of Statistical Software, 93(14)}, 1-26.
}
\author{
Wenchao Ma, The University of Minnesota, \email{wma@umn.edu}
}
\keyword{datasets}
