% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim30GDINA.R
\docType{data}
\name{sim30GDINA}
\alias{sim30GDINA}
\title{Simulated data, Q-matrix and item parameters (30 items, G-DINA model)}
\format{A list of responses, Q-matrix and item parameters for a 30-item test with components:
\describe{
\item{\code{simdat}}{simulated responses of 1000 examinees}
\item{\code{simQ}}{artificial \eqn{30 \times 5} Q-matrix}
\item{\code{simItempar}}{artificial item parameters(probability of success for each latent group)}
}}
\usage{
sim30GDINA
}
\description{
Artificial Q-matrix and item parameters for a 30-item test measuring 5 attributes
are given. Based on the Q-matrix and item parameters, responses of 1000 examinees
were simulated using the random seed of 12345.
}
\keyword{datasets}

