% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim30pGDINA.R
\docType{data}
\name{sim30pGDINA}
\alias{sim30pGDINA}
\title{Simulated data, Q-matrix and item parameters (30 items, polytomous G-DINA model)}
\format{A list of Q-matrix and item parameters for a 30-item test with components:
\describe{
\item{\code{simdat}}{simulated responses of 3000 examinees}
\item{\code{simQ}}{artificial Q-matrix}
\item{\code{simItempar}}{artificial item parameters(probability of success for each latent group)}
}}
\usage{
sim30pGDINA
}
\description{
Artificial Q-matrix and item parameters for a 30-item test measuring 5 attributes
are given. Based on the Q-matrix and item parameters, responses of 3000 examinees
were simulated using the random seed of 12345.
}
\keyword{datasets}

